Empty emplementation of the sepia filter (src/sepia.cpp) and fake test (test/sepia_test.cpp)
