+++ photoeffects/src/photoeffects.cpp
@@ -1,8 +1,1 @@
 #include "photoeffects.hpp"
-
-using namespace cv;
-
-Mat sepia(Mat& image)
-{
-    return image;
-}
