+++ photoeffects/src/sepia.cpp
@@ -1,0 +1,8 @@
+#include "photoeffects.hpp"
+
+using namespace cv;
+
+Mat sepia(Mat& image)
+{
+    return image;
+}
