+++ photoeffects/test/sepia_test.cpp
@@ -1,0 +1,11 @@
+#include <gtest/gtest.h>
+
+#include "photoeffects.hpp"
+
+using namespace cv;
+
+TEST(photoeffects, SepiaTest) {
+    Mat image(10, 10, CV_8UC1);
+
+    EXPECT_EQ(10, sepia(image).cols);
+}
